requirements.txt
requests~=2.32.3
fastapi~=0.115.11
uvicorn[standard]~=0.34.0
python-dotenv~=1.0.1